Refactors to minimize code duplication
Summary:
- Moves `leaderboard.js` to leaderboard/index.js`
- Refactors out Footer component
- Refactors out Category Leaderboard
- Refactors out Main Leaderboard
- Refactors out leaderboard categories
- Moves getWeekNumber method to utils
- LeaderboardCard shall work based on summary instead of weekSummary
