+++ lib/leaderboard.js
@@ -1,0 +1,8 @@
+export const categories = [
+  { slug: "eod_update", title: "EOD Updates" },
+  { slug: "pr_opened", title: "Pull Requests Opened" },
+  { slug: "pr_merged", title: "Pull Requests Merged" },
+  { slug: "pr_reviewed", title: "Pull Requests Reviewed" },
+  { slug: "issue_opened", title: "Issues Opened" },
+  { slug: "comment_created", title: "Comments Created" },
+];
