+++ lib/utils.js
@@ -49,3 +49,12 @@ export const parseDateString = (dateStr) => {
   dateString += date.getHours() > 11 ? " PM" : " AM";
   return dateString;
 };
+
+// Calculate week number
+export const getWeekNumber = (date) => {
+  const d = new Date(date);
+  const dayNum = d.getUTCDay() || 7;
+  d.setUTCDate(d.getUTCDate() + 4 - dayNum);
+  const yearStart = new Date(Date.UTC(d.getUTCFullYear(), 0, 1));
+  return Math.ceil(((d - yearStart) / 86400000 + 1) / 7);
+};
